Moved test files into tests/ folder
